Interupting prompts now (probably) generates event
When you interupt a prompt, a showstatus event will be generated (as
soon as the statusbar/line on the bottom of the screen appears).
This makes it possible to determin when a prompt has been closed even if
the user interupts is with ctrl+C. I doubt this is the ideal solution,
for instance, if there is no songs to play (for whatever reason) then
that statusbar may not appear, even if the prompt is closed (in which
case I need to figure out what we do then).

That being said, I believe this should work for the most part. The only
times I can think that might happen are when there is no network, no
station selected (select station prompt autopopup? if not user should
select one on there own).

Here is some basic information one would need to use the events
properly:

 - All prompt* events are requesting user input.
 - If the last event sent is a prompt event, one must respond
   appropriatly to the prompt or send ctrl+c/INT to close the prompt.
   (will generate a showstatus event IF no other events occure
   beforehand. Note all non-prompt events couunt as a status events)
 - Prompts which generate options will have those options listed as
   option1=string, option2=anotherstring, option3=... etc. A optionCount
   is also returned for all events. This is just like stationCount and
   station0,station1, etc.
 - When a prompt is opened, it appears that pianobar will not play
   anymore songs. However, it will finish the song it is currently
   playing. You can calculate the current position in the track by the
   last given songDurration and songPlayed. If songPlayed +
   timePassedSinceEventSent > songDuration then the song has probably
   finished and pianobar is waiting for you to close the prompt before
   continueing. The songfinish event will be sent as soon as the prompt
   is closed. While this may not seem  completly ideal, it is calculable
   so not realy a problem, and it is thanks to this fact that we can
   determin if prompts are open/closed so easily(if it is open, the last
   event will still be the prompt)!

Note my changes currently arn't too pretty. And even after I fix them up a bit,
there will be some difficaultis keeping in sync with upstream. I will
try to maintain this for a while, but once I find and figure out how to
use a pandora client better suited to my needs (Home audio server, im
sure there are hundreds Im just not seeing), I will reference it here
and stop maintaining this.